requirements.txt
Flask
python-docx
markdownify
flask-cors
Werkzeug
python-dotenv